Document step aggregation
