+++ lib/steps/google/assign-saml-profile.ts
@@ -1,0 +1,18 @@
+import type { StepDefinition } from "../../types";
+
+/**
+ * Step definition for assigning the Google SAML profile to users or organizational units.
+ */
+export const g7AssignSamlProfile: StepDefinition = {
+  id: "G-7",
+  title: "Assign Google SAML Profile to Users/OUs",
+  description:
+    "Activate the 'Azure AD SSO' SAML profile for users in Google Workspace. This tool assigns it to the Root OU by default. Adjust in Google Admin console if specific OUs/Groups are required.",
+  category: "SSO",
+  automatable: true,
+  requires: ["G-6"],
+  adminUrls: {
+    configure: "https://admin.google.com/ac/sso",
+    verify: "https://admin.google.com/ac/sso",
+  },
+};
