+++ lib/steps/microsoft/start-provisioning.ts
@@ -1,0 +1,29 @@
+import type { StepDefinition } from "../../types";
+import { OUTPUT_KEYS } from "../../types";
+/**
+ * Step definition for starting the Azure AD provisioning job.
+ */
+
+export const m5StartProvisioning: StepDefinition = {
+  id: "M-5",
+  title: "Define Scope & Start Provisioning Job",
+  description:
+    "This action starts the Azure AD provisioning job. IMPORTANT: Before starting, manually configure the provisioning scope (which users/groups to sync) in the Azure Portal for the 'Google Workspace User Provisioning' app.",
+  category: "Microsoft",
+  automatable: true,
+  requires: ["M-4"],
+  adminUrls: {
+    configure: (outputs) => {
+      const spId = outputs[OUTPUT_KEYS.PROVISIONING_SP_OBJECT_ID];
+      const appId = outputs[OUTPUT_KEYS.PROVISIONING_APP_ID];
+      if (!spId || !appId) return null;
+      return `https://portal.azure.com/#view/Microsoft_AAD_IAM/ManagedAppMenuBlade/~/ProvisioningManagement/appId/${appId}/objectId/${spId}`;
+    },
+    verify: (outputs) => {
+      const spId = outputs[OUTPUT_KEYS.PROVISIONING_SP_OBJECT_ID];
+      const appId = outputs[OUTPUT_KEYS.PROVISIONING_APP_ID];
+      if (!spId || !appId) return null;
+      return `https://portal.azure.com/#view/Microsoft_AAD_IAM/ManagedAppMenuBlade/~/ProvisioningManagement/appId/${appId}/objectId/${spId}`;
+    },
+  },
+};
